+++ code/arguments.py
@@ -75,7 +75,14 @@ class DataTrainingArguments:
         default=64, metadata={"help": "Define how many clusters to use for faiss."}
     )
     top_k_retrieval: int = field(
-        default=10,
+        default=30,
         metadata={"help": "Define how many top-k passages to retrieve based on similarity."},
     )
-    use_faiss: bool = field(default=False, metadata={"help": "Whether to build with faiss"})
+    use_faiss: bool = field(
+        default=False, 
+        metadata={"help": "Whether to build with faiss"}
+    )
+    bm25: bool = field(
+        default=True, 
+        metadata={"help": "Whether to use BM25"}
+    )
